fix task-hot-swap by moving restart method into api (#1858)
